+++ src/ReachSets/ContinuousPost/options_blocks.jl
@@ -1,194 +1,3 @@
-import Base: merge, merge!, getindex, keys, haskey, values, setindex!, copy
-
-export Options, merge, merge!, getindex, haskey
-
-available_keywords = Set{Symbol}([])
-
-"""
-    Options
-
-Type that wraps a dictionary used for options.
-
-### Fields
-
-- `dict` -- the wrapped dictionary
-"""
-struct Options
-    dict::Dict{Symbol,Any}
-    Options(args::Pair{Symbol,<:Any}...) = new(Dict{Symbol,Any}(args))
-    Options(dict::Dict{Symbol,Any}) = new(dict)
-end
-
-"""
-    keys(op::Options)
-
-Return the keys of the given options object.
-
-### Input
-
-- `op`    -- options object
-
-### Examples
-
-Obtain the keys of some options with one element:
-
-```jldoctest options_setindex
-julia> op = Options(:T=>1.0)
-Reachability.Options(Dict{Symbol,Any}(Pair{Symbol,Any}(:T, 1.0)))
-
-julia> collect(keys(op))
-1-element Array{Symbol,1}:
- :T
-```
-"""
-keys(op::Options) = keys(op.dict)
-
-"""
-    values(op::Options)
-
-Return the values of the given options object.
-
-### Input
-
-- `op`    -- options object
-
-### Examples
-
-Obtain the values of some options with one element:
-
-```jldoctest options_setindex
-julia> op = Options(:T=>1.0)
-Reachability.Options(Dict{Symbol,Any}(Pair{Symbol,Any}(:T, 1.0)))
-
-julia> collect(values(op))
-1-element Array{Any,1}:
- 1.0
-```
-"""
-values(op::Options) = values(op.dict)
-
-"""
-    merge(op1, opn)
-
-Merges two `Options` objects by just falling back to the wrapped `Dict` fields.
-Values are inserted in the order in which the function arguments occur, i.e.,
-for conflicting keys a later object overrides a previous value.
-
-### Input
-
-- `op1` -- first options object
-- `opn` -- list of options objects
-
-### Output
-
-An `Options` object.
-
-### Notes
-
-This function makes a copy of the dictionary and does not modify its first
-argument.
-"""
-function merge(op1::Options, opn::Options...)::Options
-    merge!(copy(op1), opn...)
-end
-
-"""
-    merge!(op1, opn)
-
-Updates the first argument options `op1` with options from `opn`.
-
-### Input
-
-- `op1` -- first options object
-- `opn` -- list of options objects
-
-### Output
-
-An `Options` object.
-
-### Algorithm 
-
-Merges two `Options` objects by just falling back to the wrapped `Dict` fields.
-Values are inserted in the order in which the function arguments occur, i.e.,
-for conflicting keys a later object overrides a previous value.
-"""
-function merge!(op1::Options, opn::Options...)::Options
-    dict = op1.dict
-    for i in 1 : length(opn)
-        merge!(dict, opn[i].dict)
-    end
-    return Options(dict)
-end
-
-"""
-    getindex(op::Options, sym::Symbol)
-
-Returns the value stored for key `sym`.
-
-### Input
-
-- `op`  -- options object
-- `sym` -- key
-"""
-function getindex(op::Options, sym::Symbol)
-    return getindex(op.dict, sym)
-end
-
-"""
-    setindex!(op, value, key)
-
-Store the given value at the given key in the options.
-
-### Input
-
-- `op`    -- options object
-- `value` -- value
-- `key`   -- key
-
-### Examples
-
-Create an empty options object and add an input:
-
-```jldoctest options_setindex
-julia> Options()
-Reachability.Options(Dict{Symbol,Any}())
-
-julia> op[:T] = 1.0
-1.0
-```
-"""
-setindex!(op::Options, value, key) = setindex!(op.dict, value, key)
-
-"""
-    copy(op)
-
-Create a shallow copy of the given options.
-
-### Input
-
-- `op`    -- options object
-
-### Output
-
-A new `Options` instance whose dictionary is a copy of `op`'s dictionary.
-"""
-copy(op::Options) = Options(copy(op.dict))
-
-"""
-    haskey(op::Options, key)
-
-Determine whether the given options has a mapping for a given key.
-
-### Input
-
-- `op` -- options object
-
-### Output
-
-`true` if `op` contains the option `key` and `false` otherwise.
-"""
-haskey(op::Options, key) = haskey(op.dict, key)
-
 """
     validate_solver_options_and_add_default_values!(options)
 
@@ -485,17 +294,22 @@ function validate_solver_options_and_add_default_values!(options::Options)::Opti
 end
 
 """
-Handling of the special trio `:δ`, `:N`, `:T`.
-Usually two of them should be defined and the third one is automatically inferred.
+    check_and_add_δ_N_T!(dict::Dict{Symbol,Any}, dict_copy::Dict{Symbol,Any})
+
+Handling of the special triple `:δ`, `:N`, `:T`. Usually two of them should be
+defined and the third one is automatically inferred.
 If three of them are defined, they must be consistent.
 If only `:T` is defined, we use `:N = 100`.
 
-INPUT:
+### Input
 
 - `dict`      -- a dictionary of options
 - `dict_copy` -- a copy of the dictionary of options for internal names
 
+### Notes
+
 Supported options:
+
 - `:δ` (time step)
 - alias: `:sampling_time`
 - `:N` (number of time steps)
@@ -814,111 +628,4 @@ function check_and_add_lazy_inputs_interval!(dict::Dict{Symbol,Any},
     else
         dict_copy[:lazy_inputs_interval] = nothing
     end
-end
-
-"""
-    check_aliases!(dict, dict_copy, aliases)
-
-This function has several purposes:
-
-- translate aliases to the option that is used internally
-- check that not several aliases were used at the same time
-
-INPUT:
-
-- `dict` -- a dictionary of options
-- `dict_copy` -- a copy of the dictionary of options for internal names
-- `aliases` -- option aliases; the first name is the one we use internally
-"""
-function check_aliases!(dict::Dict{Symbol,Any}, dict_copy::Dict{Symbol,Any}, aliases::Vector{Symbol})
-    # find aliases and check consistency in case several aliases have been used
-    print_warning::Bool = false
-    for alias in aliases
-        # add alias to global keyword set
-        push!(available_keywords, alias)
-
-        if haskey(dict, alias)
-            # alias was used
-            if haskey(dict_copy, aliases[1])
-                # several aliases were used
-                if (dict_copy[aliases[1]] != dict[alias])
-                    error("Several option aliases were used with different values for aliases $aliases.")
-                else
-                    print_warning = true
-                end
-            else
-                dict_copy[aliases[1]] = dict[alias]
-            end
-        end
-    end
-    if print_warning
-        warn("Several option aliases were used for aliases $aliases.")
-    end
-end
-
-"""
-    check_aliases_and_add_default_value!(dict, dict_copy, aliases, default_value, [modify_dict])
-
-This function has several purposes:
-
-- translate aliases to the option that is used internally
-- check that not several aliases were used at the same time
-- assign default values for undefined options
-
-### Input
-
-- `dict`          -- a dictionary of options
-- `dict_copy`     -- a copy of the dictionary of options for internal names
-- `aliases`       -- option aliases; the first name is the one we use internally
-- `default_value` -- the default value for the option
-- `modify_dict`   -- (optional, default: `false`) indicates if `dict` should be
-                     modified
-"""
-function check_aliases_and_add_default_value!(dict::Dict{Symbol,Any}, dict_copy::Dict{Symbol,Any}, aliases::Vector{Symbol}, default_value::Any, modify_dict::Bool=false)
-    check_aliases!(dict, dict_copy, aliases)
-
-    if !haskey(dict_copy, aliases[1])
-        # no alias and no value
-        # TODO<notification> print message to user for each default option, depending on verbosity level
-        if (modify_dict)
-            dict[aliases[1]] = default_value
-        end
-        dict_copy[aliases[1]] = default_value
-    end
-end
-
-"""
-    check_valid_option_keywords(dict)
-
-Validate that all input keywords are recognized.
-
-### Input
-
-- `dict` -- input dictionary
-"""
-function check_valid_option_keywords(dict)
-    for kv_pair in dict
-        key::Symbol = kv_pair.first
-        if !in(key, available_keywords)
-            error(get_unrecognized_key_message(key))
-        end
-    end
-end
-
-"""
-    get_unrecognized_key_message(key)
-
-Create an error message for an unrecognized option key.
-
-### Input
-
-- `key` -- unrecognized option key
-
-### Output
-
-The error message.
-"""
-function get_unrecognized_key_message(key)
-    return "Unrecognized option '$key' found. See " *
-        "`keys(Reachability.available_keywords.dict)` for all valid keywords."
 end
